+++ src/apiCalls/account/getAccountFromApi.ts
@@ -4,8 +4,14 @@ import { getCleanApiAddress } from 'apiCalls/utils/getCleanApiAddress';
 import { TIMEOUT } from 'constants/network.constants';
 import { AccountType } from 'types/account.types';
 
-export const accountFetcher = (address: string | null) => {
-  const apiAddress = getCleanApiAddress();
+export const accountFetcher = ({
+  address,
+  baseURL
+}: {
+  address: string | null;
+  baseURL: string;
+}) => {
+  const apiAddress = getCleanApiAddress(baseURL);
   const url = `${apiAddress}/${ACCOUNTS_ENDPOINT}/${address}?withGuardianInfo=true`;
   // we need to get it with an axios instance because of cross-window user interaction issues
   return axiosInstance.get(url, {
@@ -14,13 +20,19 @@ export const accountFetcher = (address: string | null) => {
   });
 };
 
-export const getAccountFromApi = async (address?: string) => {
+export const getAccountFromApi = async ({
+  address,
+  baseURL
+}: {
+  address?: string;
+  baseURL: string;
+}) => {
   if (!address) {
     return null;
   }
 
   try {
-    const { data } = await accountFetcher(address);
+    const { data } = await accountFetcher({ address, baseURL });
     return data as AccountType;
   } catch (_err) {
     console.error('error fetching configuration for ', address);