+++ src/core/providers/strategies/helpers/signTransactions/signTransactions.ts
@@ -1,35 +1,24 @@
 import { Transaction } from '@multiversx/sdk-core/out';
 import { getEconomics } from 'apiCalls/economics/getEconomics';
-import { getPersistedTokenDetails } from 'apiCalls/tokens/getPersistedTokenDetails';
-import {
-  GAS_PER_DATA_BYTE,
-  GAS_PRICE_MODIFIER,
-  MULTI_TRANSFER_EGLD_TOKEN
-} from 'constants/mvx.constants';
+import { EMPTY_PPU } from 'constants/placeholders.constants';
 import { UITagsEnum } from 'constants/UITags.enum';
 import { SignTransactionsStateManager } from 'core/managers/internal/SignTransactionsStateManager/SignTransactionsStateManager';
 import {
-  ISignTransactionsModalData,
+  ISignTransactionsModalCommonData,
   SignEventsEnum
 } from 'core/managers/internal/SignTransactionsStateManager/types';
-import { getAddress } from 'core/methods/account/getAddress';
+import { getAccountInfo } from 'core/methods/account/getAccountInfo';
 import { getEgldLabel } from 'core/methods/network/getEgldLabel';
 import { cancelCrossWindowAction } from 'core/providers/helpers/cancelCrossWindowAction';
 import { IProvider } from 'core/providers/types/providerFactory.types';
 import { SignTransactionsModal } from 'lib/sdkDappCoreUi';
-import { formatAmount } from 'lib/sdkDappUtils';
 import { networkSelector } from 'store/selectors/networkSelectors';
 import { getState } from 'store/store';
-import { NftEnumType } from 'types/tokens.types';
 import { createUIElement } from 'utils/createUIElement';
-import { calculateFeeInFiat } from './helpers/calculateFeeInFiat';
-import { calculateFeeLimit } from './helpers/calculateFeeLimit';
-import { getExtractTransactionsInfo } from './helpers/getExtractTransactionsInfo';
-import { getHighlight } from './helpers/getHighlight';
+import { getCommonData } from './helpers/getCommonData/getCommonData';
+import { getRecommendedGasPrice } from './helpers/getCommonData/helpers/getRecommendedGasPrice';
+import { getFeeData } from './helpers/getFeeData';
 import { getMultiEsdtTransferData } from './helpers/getMultiEsdtTransferData/getMultiEsdtTransferData';
-import { getScCall } from './helpers/getScCall';
-import { getTokenType } from './helpers/getTokenType';
-import { getUsdValue } from './helpers/getUsdValue';
 import { guardTransactions as getGuardedTransactions } from './helpers/guardTransactions/guardTransactions';
 
 type SignTransactionsParamsType = {
@@ -43,7 +32,9 @@ export async function signTransactions({
   handleSign,
   guardTransactions = getGuardedTransactions
 }: SignTransactionsParamsType): Promise<Transaction[]> {
-  const address = getAddress();
+  const {
+    account: { address, shard }
+  } = getAccountInfo();
   const network = networkSelector(getState());
 
   const egldLabel = getEgldLabel();
@@ -51,8 +42,10 @@ export async function signTransactions({
     name: UITagsEnum.SIGN_TRANSACTIONS_MODAL
   });
 
-  const { allTransactions, getTxInfoByDataField } =
+  const { allTransactions, parsedTransactionsByDataField } =
     getMultiEsdtTransferData(transactions);
+
+  let signedIndexes: number[] = [];
 
   const eventBus = await signModalElement.getEventBus();
 
@@ -67,131 +60,79 @@ export async function signTransactions({
 
   return new Promise<Transaction[]>(async (resolve, reject) => {
     const signedTransactions: Transaction[] = [];
-    let currentTransactionIndex = 0;
-    const economics = await getEconomics();
+    const economics = await getEconomics({ baseURL: network.apiAddress });
 
-    const signNextTransaction = async () => {
-      const currentTransaction = allTransactions[currentTransactionIndex];
-      const sender = currentTransaction?.transaction?.getSender().toString();
+    manager.initializeGasPriceMap(allTransactions.map((tx) => tx.transaction));
+
+    const showNextScreen = async (currentScreenIndex: number) => {
+      const currentTransaction = allTransactions[currentScreenIndex];
       const transaction = currentTransaction?.transaction;
       const price = economics?.price;
+      const currentNonce = currentTransaction.transaction?.getNonce().valueOf();
 
-      const feeLimit = calculateFeeLimit({
-        gasPerDataByte: String(GAS_PER_DATA_BYTE),
-        gasPriceModifier: String(GAS_PRICE_MODIFIER),
-        gasLimit: transaction.getGasLimit().valueOf().toString(),
-        gasPrice: transaction.getGasPrice().valueOf().toString(),
-        data: transaction.getData().toString(),
-        chainId: transaction.getChainID().valueOf()
-      });
-
-      const feeLimitFormatted = formatAmount({
-        input: feeLimit,
-        showLastNonZeroDecimal: true
-      });
-
-      const feeInFiatLimit = price
-        ? calculateFeeInFiat({
-            feeLimit,
-            egldPriceInUsd: price,
-            hideEqualSign: true
-          })
-        : null;
-
-      const extractTransactionsInfo = getExtractTransactionsInfo({
-        getTxInfoByDataField,
-        egldLabel,
-        sender,
-        address
-      });
-
-      const plainTransaction = currentTransaction.transaction.toPlainObject();
-
-      const txInfo = await extractTransactionsInfo(currentTransaction);
-
-      const isEgld = !txInfo?.transactionTokenInfo?.tokenId;
-      const {
-        tokenId,
-        nonce,
-        amount = ''
-      } = txInfo?.transactionTokenInfo ?? {};
-
-      const isNftOrSft = tokenId && nonce && nonce.length > 0;
-      const tokenIdForTokenDetails = isNftOrSft
-        ? `${tokenId}-${nonce}`
-        : tokenId;
-
-      const tokenDetails = await getPersistedTokenDetails({
-        tokenId: tokenIdForTokenDetails
-      });
-
-      const { esdtPrice, tokenDecimals, type, identifier, tokenImageUrl } =
-        tokenDetails;
-
-      const isNft =
-        type === NftEnumType.SemiFungibleESDT ||
-        type === NftEnumType.NonFungibleESDT;
-
-      if (isNft) {
-        manager.updateNonFungibleTransaction(type, {
-          identifier,
-          amount,
-          imageURL: tokenImageUrl
-        });
-      } else {
-        const getFormattedAmount = ({ addCommas }: { addCommas: boolean }) =>
-          formatAmount({
-            input: isEgld
-              ? currentTransaction.transaction.getValue().toString()
-              : amount,
-            decimals: isEgld ? Number(network.decimals) : tokenDecimals,
-            digits: Number(network.digits),
-            showLastNonZeroDecimal: false,
-            addCommas
-          });
-
-        const formattedAmount = getFormattedAmount({ addCommas: true });
-        const rawAmount = getFormattedAmount({ addCommas: false });
-        const tokenPrice = Number(isEgld ? price : esdtPrice);
-        const usdValue = getUsdValue({
-          amount: rawAmount,
-          usd: tokenPrice,
-          addEqualSign: true
+      const { commonData, tokenTransaction, fungibleTransaction } =
+        await getCommonData({
+          allTransactions,
+          currentScreenIndex,
+          egldLabel,
+          network,
+          gasPriceData: manager.ppuMap[currentNonce],
+          price,
+          address,
+          shard,
+          signedIndexes,
+          parsedTransactionsByDataField
         });
 
-        const esdtIdentifier =
-          identifier === MULTI_TRANSFER_EGLD_TOKEN ? egldLabel : identifier;
-        manager.updateTokenTransaction({
-          identifier: esdtIdentifier ?? egldLabel,
-          amount: formattedAmount,
-          usdValue
-        });
+      if (tokenTransaction) {
+        manager.updateTokenTransaction(tokenTransaction);
       }
 
-      const commonData: ISignTransactionsModalData['commonData'] = {
-        receiver: plainTransaction.receiver.toString(),
-        data: currentTransaction.transaction.getData().toString(),
-        egldLabel,
-        tokenType: getTokenType(type),
-        feeLimit: feeLimitFormatted,
-        feeInFiatLimit,
-        transactionsCount: allTransactions.length,
-        currentIndex: currentTransactionIndex,
-        highlight: getHighlight(txInfo?.transactionTokenInfo),
-        scCall: getScCall(txInfo?.transactionTokenInfo)
-      };
+      if (fungibleTransaction) {
+        manager.updateNonFungibleTransaction(
+          fungibleTransaction.type,
+          fungibleTransaction
+        );
+      }
 
       manager.updateCommonData(commonData);
-      manager.updateConfirmedTransactions();
 
-      const onPreviousTransaction = async () => {
-        const data = manager.confirmedScreens[manager.currentScreenIndex - 1];
-        manager.updateData(data);
+      const onBack = () => {
+        removeEvents();
+        showNextScreen(currentScreenIndex - 1);
       };
 
-      const onNextTransaction = async () => {
-        const data = manager.confirmedScreens[manager.currentScreenIndex + 1];
-        manager.updateData(data);
+      const onSetPpu = (
+        ppu: ISignTransactionsModalCommonData['ppu'] = EMPTY_PPU
+      ) => {
+        manager.updateGasPriceMap({
+          nonce: currentNonce,
+          ppu
+        });
+
+        const plainTransaction = transaction.toPlainObject();
+
+        const newGasPrice = getRecommendedGasPrice({
+          transaction: plainTransaction,
+          gasPriceData: manager.ppuMap[currentNonce]
+        });
+
+        const newTransaction = Transaction.fromPlainObject({
+          ...plainTransaction,
+          gasPrice: newGasPrice
+        });
+
+        const feeData = getFeeData({
+          transaction: newTransaction,
+          price
+        });
+
+        manager.updateCommonData({
+          feeLimit: feeData.feeLimitFormatted,
+          feeInFiatLimit: feeData.feeInFiatLimit,
+          gasPrice: newGasPrice.toString(),
+          ppu
+        });
       };
 
       const onCancel = async () => {
@@ -200,49 +141,55 @@ export async function signTransactions({
         signModalElement.remove();
       };
 
-      const onSign = async () => {
-        const shouldContinueWithoutSigning = Boolean(
-          txInfo?.transactionTokenInfo?.type &&
-            txInfo?.transactionTokenInfo?.multiTxData &&
-            !txInfo?.dataField.endsWith(
-              txInfo?.transactionTokenInfo?.multiTxData
-            )
-        );
+      function removeEvents() {
+        eventBus.unsubscribe(SignEventsEnum.CONFIRM, onSign);
+        eventBus.unsubscribe(SignEventsEnum.CLOSE, onCancel);
+        eventBus.unsubscribe(SignEventsEnum.BACK, onBack);
+        eventBus.unsubscribe(SignEventsEnum.SET_PPU, onSetPpu);
+      }
 
-        const removeEvents = () => {
-          eventBus.unsubscribe(SignEventsEnum.SIGN_TRANSACTION, onSign);
-          eventBus.unsubscribe(SignEventsEnum.CLOSE, onCancel);
-          eventBus.unsubscribe(
-            SignEventsEnum.PREV_TRANSACTION,
-            onPreviousTransaction
-          );
-          eventBus.unsubscribe(
-            SignEventsEnum.NEXT_TRANSACTION,
-            onNextTransaction
-          );
-        };
+      async function onSign() {
+        const shouldContinueWithoutSigning = !commonData.needsSigning;
+
+        removeEvents();
 
         if (shouldContinueWithoutSigning) {
-          currentTransactionIndex++;
-          removeEvents();
-          manager.setNextUnsignedTxIndex(currentTransactionIndex);
-          return signNextTransaction();
+          return showNextScreen(currentScreenIndex + 1);
         }
 
+        const currentEditedTransaction = currentTransaction.transaction;
+
+        const txNonce = currentEditedTransaction.getNonce().valueOf();
+
+        if (!currentNonce) {
+          throw new Error('Current nonce not found');
+        }
+
+        const newGasPrice = getRecommendedGasPrice({
+          transaction: currentEditedTransaction.toPlainObject(),
+          gasPriceData: manager.ppuMap[txNonce]
+        });
+
+        const transactionToSign = Transaction.fromPlainObject({
+          ...currentEditedTransaction.toPlainObject(),
+          gasPrice: newGasPrice
+        });
+
         try {
-          const signedTransaction = await handleSign([
-            currentTransaction.transaction
-          ]);
+          const signedTransaction = await handleSign([transactionToSign]);
 
           if (signedTransaction) {
+            signedIndexes.push(currentScreenIndex);
             signedTransactions.push(signedTransaction[0]);
           }
 
-          removeEvents();
+          const isLastScreen =
+            currentScreenIndex === allTransactions.length - 1;
 
-          const areAllSigned = signedTransactions.length == transactions.length;
+          const areAllTransactionsSigned =
+            signedTransactions.length === transactions.length;
 
-          if (areAllSigned) {
+          if (isLastScreen && areAllTransactionsSigned) {
             const optionallyGuardedTransactions =
               await guardTransactions(signedTransactions);
             signModalElement.remove();
@@ -250,24 +197,19 @@ export async function signTransactions({
             return resolve(optionallyGuardedTransactions);
           }
 
-          currentTransactionIndex++;
-          manager.setNextUnsignedTxIndex(currentTransactionIndex);
-          signNextTransaction();
+          showNextScreen(currentScreenIndex + 1);
         } catch (error) {
           reject('Error signing transactions: ' + error);
           signModalElement.remove();
         }
-      };
+      }
 
-      eventBus.subscribe(SignEventsEnum.SIGN_TRANSACTION, onSign);
+      eventBus.subscribe(SignEventsEnum.CONFIRM, onSign);
       eventBus.subscribe(SignEventsEnum.CLOSE, onCancel);
-      eventBus.subscribe(
-        SignEventsEnum.PREV_TRANSACTION,
-        onPreviousTransaction
-      );
-      eventBus.subscribe(SignEventsEnum.NEXT_TRANSACTION, onNextTransaction);
+      eventBus.subscribe(SignEventsEnum.BACK, onBack);
+      eventBus.subscribe(SignEventsEnum.SET_PPU, onSetPpu);
     };
 
-    signNextTransaction();
+    showNextScreen(0);
   });
 }