+++ src/lib/sdkDappCoreUi.ts
@@ -1,11 +1,11 @@
-import type { CustomElementsDefineOptions } from '@multiversx/sdk-dapp-core-ui/loader';
-export type { LedgerConnectModal } from '@multiversx/sdk-dapp-core-ui/dist/components/ledger-connect-modal';
-export type { LedgerConnect } from '@multiversx/sdk-dapp-core-ui/dist/components/ledger-connect';
-export type { SignTransactionsModal } from '@multiversx/sdk-dapp-core-ui/dist/components/sign-transactions-modal';
-export type { WalletConnectModal } from '@multiversx/sdk-dapp-core-ui/dist/components/wallet-connect-modal';
-export type { PendingTransactionsModal } from '@multiversx/sdk-dapp-core-ui/dist/components/pending-transactions-modal';
-export type { NotificationsFeed } from '@multiversx/sdk-dapp-core-ui/dist/components/notifications-feed';
-export type { ToastList } from '@multiversx/sdk-dapp-core-ui/dist/components/toast-list';
+import type { CustomElementsDefineOptions } from '@multiversx/sdk-dapp-core-ui/dist/loader';
+export type { LedgerConnectModal } from '@multiversx/sdk-dapp-core-ui/dist/web-components/ledger-connect-modal';
+export type { LedgerConnect } from '@multiversx/sdk-dapp-core-ui/dist/web-components/ledger-connect';
+export type { SignTransactionsModal } from '@multiversx/sdk-dapp-core-ui/dist/web-components/sign-transactions-modal';
+export type { WalletConnectModal } from '@multiversx/sdk-dapp-core-ui/dist/web-components/wallet-connect-modal';
+export type { PendingTransactionsModal } from '@multiversx/sdk-dapp-core-ui/dist/web-components/pending-transactions-modal';
+export type { NotificationsFeed } from '@multiversx/sdk-dapp-core-ui/dist/web-components/notifications-feed';
+export type { ToastList } from '@multiversx/sdk-dapp-core-ui/dist/web-components/toast-list';
 export type { IEventBus } from '@multiversx/sdk-dapp-core-ui/dist/types/utils/EventBus';
 export type {
   ITransactionListItem,
@@ -19,7 +19,7 @@ export async function defineCustomElements(
   opts?: CustomElementsDefineOptions
 ): Promise<void> {
   try {
-    const loader = await import('@multiversx/sdk-dapp-core-ui/loader');
+    const loader = await import('@multiversx/sdk-dapp-core-ui/dist/loader');
     loader.defineCustomElements(win, opts);
   } catch (err) {
     throw new Error('@multiversx/sdk-dapp-core-ui not found' + err);